STREAMLINE-708: Address review comments
1. Added case insensitive matching
2. Return enriched entities when detail flag is set in the query params
